+++ core/src/elements/ons-button.js
@@ -49,6 +49,9 @@ const scheme = {'': 'button--*'};
  * @modifier material
  *   [en]Material Design button[/en]
  *   [ja]マテリアルデザインのボタン[/ja]
+ * @modifier material--flat
+ *   [en]Material Design flat button[/en]
+ *   [ja]マテリアルデザインのフラットボタン[/ja]
  * @description
  *   [en]
  *     Button component. If you want to place a button in a toolbar, use `<ons-toolbar-button>` or `<ons-back-button>` instead.
@@ -57,6 +60,7 @@ const scheme = {'': 'button--*'};
  *   [/en]
  *   [ja]ボタン用コンポーネント。ツールバーにボタンを設置する場合は、ons-toolbar-buttonもしくはons-back-buttonコンポーネントを使用します。[/ja]
  * @codepen hLayx
+ * @tutorial vanilla/Reference/button
  * @guide Button [en]Guide for `<ons-button>`[/en][ja]<ons-button>の使い方[/ja]
  * @guide OverridingCSSstyles [en]More details about the `modifier` attribute[/en][ja]modifier属性の使い方[/ja]
  * @example