+++ core/src/elements/ons-input.js
@@ -65,10 +65,11 @@ const INPUT_ATTRIBUTES = [
  *
  *    The component will automatically render as a Material Design input on Android devices.
  *
- *    Most attributes that can be used for a normal `<input>` element can also be used on the `<ons-input>` element..
+ *    Most attributes that can be used for a normal `<input>` element can also be used on the `<ons-input>` element.
  *  [/en]
  *  [ja][/ja]
  * @codepen ojQxLj
+ * @tutorial vanilla/Reference/input
  * @seealso ons-range
  *   [en]The `<ons-range>` element is used to display a range slider.[/en]
  *   [ja][/ja]
@@ -118,7 +119,7 @@ class InputElement extends BaseElement {
    * @attribute input-id
    * @type {String}
    * @description
-   *  [en]Specify the "id" attribute of the inner <input> element. This is useful when using <label for="..."> elements.[/en]
+   *  [en]Specify the "id" attribute of the inner `<input>` element. This is useful when using <label for="..."> elements.[/en]
    *  [ja][/ja]
    */
 
@@ -199,7 +200,7 @@ class InputElement extends BaseElement {
 
   attributeChangedCallback(name, last, current) {
     if (name === 'modifier') {
-      return ModifierUtil.onModifierChanged(last, current, this, scheme);
+      return contentReady(this, () => ModifierUtil.onModifierChanged(last, current, this, scheme));
     } else if (name === 'placeholder') {
       return contentReady(this, () => this._updateLabel());
     } if (name === 'input-id') {
@@ -213,22 +214,26 @@ class InputElement extends BaseElement {
   }
 
   attachedCallback() {
-    if (this.type !== 'checkbox' && this.type !== 'radio') {
-      this.addEventListener('input', this._boundOnInput);
-      this.addEventListener('focusin', this._boundOnFocusin);
-      this.addEventListener('focusout', this._boundOnFocusout);
-    }
+    contentReady(this, () => {
+      if (this._input.type !== 'checkbox' && this._input.type !== 'radio') {
+        this._input.addEventListener('input', this._boundOnInput);
+        this._input.addEventListener('focusin', this._boundOnFocusin);
+        this._input.addEventListener('focusout', this._boundOnFocusout);
+      }
 
-    this.addEventListener('focus', this._boundDelegateEvent);
-    this.addEventListener('blur', this._boundDelegateEvent);
+      this._input.addEventListener('focus', this._boundDelegateEvent);
+      this._input.addEventListener('blur', this._boundDelegateEvent);
+    });
   }
 
   detachedCallback() {
-    this.removeEventListener('input', this._boundOnInput);
-    this.removeEventListener('focusin', this._boundOnFocusin);
-    this.removeEventListener('focusout', this._boundOnFocusout);
-    this.removeEventListener('focus', this._boundDelegateEvent);
-    this.removeEventListener('blur', this._boundDelegateEvent);
+    contentReady(this, () => {
+      this._input.removeEventListener('input', this._boundOnInput);
+      this._input.removeEventListener('focusin', this._boundOnFocusin);
+      this._input.removeEventListener('focusout', this._boundOnFocusout);
+      this._input.removeEventListener('focus', this._boundDelegateEvent);
+      this._input.removeEventListener('blur', this._boundDelegateEvent);
+    });
   }
 
   _setLabel(value) {