+++ core/src/elements/ons-modal/index.spec.js
@@ -108,28 +108,20 @@ describe('OnsModalElement', () => {
     });
   });
 
-  describe('#_onDeviceBackButton()', () => {
-    it('returns nothing by default', () => {
-      expect(element._onDeviceBackButton()).not.to.be.ok;
-    });
-  });
-
   describe('#onDeviceBackButton', () => {
-    it('gets the callback', (done) => {
-      setTimeout(() => {
-        expect(element.onDeviceBackButton).to.be.ok;
-        done();
-      }, 10);
+    it('gets the callback', () => {
+      expect(element.onDeviceBackButton).to.be.ok;
     });
 
-    it('overwrites the callback', (done) => {
-      setTimeout(() => {
-        const spy = chai.spy.on(element._backButtonHandler, 'destroy');
-        element.onDeviceBackButton = () => { return; };
-        expect(spy).to.have.been.called.once;
-        expect(element._backButtonHandler).to.be.ok;
-        done();
-      }, 10);
+    it('returns nothing by default', () => {
+      expect(element.onDeviceBackButton._callback()).not.to.be.ok;
+    });
+
+    it('overwrites the callback', () => {
+      const spy = chai.spy.on(element._backButtonHandler, 'destroy');
+      element.onDeviceBackButton = () => { return; };
+      expect(spy).to.have.been.called.once;
+      expect(element._backButtonHandler).to.be.ok;
     });
   });