+++ core/src/elements/ons-pull-hook.js
@@ -34,6 +34,7 @@ const STATE_ACTION = 'action';
  *   [/en]
  *   [ja][/ja]
  * @codepen WbJogM
+ * @tutorial vanilla/Reference/pull-hook
  * @guide UsingPullHook
  *   [en]How to use Pull Hook[/en]
  *   [ja]プルフックを使う[/ja]
@@ -106,6 +107,8 @@ class PullHookElement extends BaseElement {
 
     this._currentTranslation = 0;
 
+    this._ensureScrollElement();
+
     this._setState(STATE_INITIAL, true);
     this._setStyle();
   }
@@ -126,6 +129,12 @@ class PullHookElement extends BaseElement {
     pageElement.appendChild(scrollElement);
 
     return scrollElement;
+  }
+
+  _ensureScrollElement() {
+    if (!this._scrollElement) {
+      this._scrollElement = this._createScrollElement();
+    }
   }
 
   _setStyle() {
@@ -429,7 +438,7 @@ class PullHookElement extends BaseElement {
   }
 
   attachedCallback() {
-    this._scrollElement = this._createScrollElement();
+    this._ensureScrollElement();
 
     this._pageElement = this._scrollElement.parentElement;