+++ core/src/elements/ons-speed-dial-item.js
@@ -31,6 +31,7 @@ const scheme = {
  *     Material DesignのSpeed dialの子要素を表現する要素です。
  *   [/ja]
  * @codepen dYQYLg
+ * @tutorial vanilla/Reference/speed-dial
  * @seealso ons-speed-dial
  *   [en]The `<ons-speed-dial>` component.[/en]
  *   [ja]ons-speed-dialコンポーネント[/ja]
@@ -55,9 +56,7 @@ class SpeedDialItemElement extends BaseElement {
    */
 
   createdCallback() {
-    if (!this.hasAttribute('_compiled')) {
-      this._compile();
-    }
+    this._compile();
 
     this._boundOnClick = this._onClick.bind(this);
   }
@@ -98,8 +97,6 @@ class SpeedDialItemElement extends BaseElement {
     this._updateRipple();
 
     ModifierUtil.initModifier(this, scheme);
-
-    this.setAttribute('_compiled', '');
   }
 }