+++ core/src/elements/ons-splitter-side.js
@@ -24,6 +24,8 @@ import BaseElement from 'ons/base-element';
 import animatorFactory from './ons-splitter/animator';
 import GestureDetector from 'ons/gesture-detector';
 import DoorLock from 'ons/doorlock';
+import contentReady from 'ons/content-ready';
+import OnsSplitterElement from './ons-splitter';
 
 const SPLIT_MODE = 'split';
 const COLLAPSE_MODE = 'collapse';
@@ -230,6 +232,7 @@ class CollapseMode {
  *  [/en]
  *  [ja]ons-splitter-side要素は、ons-splitter要素の子要素として利用します。[/ja]
  * @codepen rOQOML
+ * @tutorial vanilla/Reference/splitter
  * @seealso ons-splitter
  *  [en]The `<ons-splitter>` is the parent component.[/en]
  *  [ja]ons-splitterコンポーネント[/ja]
@@ -436,7 +439,10 @@ class SplitterSideElement extends BaseElement {
     if (!this.hasAttribute('side')) {
       this.setAttribute('side', 'left');
     }
-    this._watchedAttributes.forEach(e => this._update(e));
+
+    contentReady(this, () => {
+      this._watchedAttributes.forEach(e => this._update(e));
+    });
   }
 
   detachedCallback() {